createOrderSpecAccount
Hey, great library. I've added a small extension to the createOrder function, that allows you to specify a specific account that you'd like the order to route to. If you're trading under an FA account or any account with multiple users/accounts the default way the function is written will throw an error. 

Cheers, 

Tyler